refactor(github): modularize GitHubService into feature mixins and re-export unified class
- Create feature mixins: http, repos, branches, prs, graphql, microagents, installations
- Assemble in openhands/integrations/github/service/__init__.py
- Add GitHubMixinBase for typing to satisfy mypy
- Preserve get_impl() dynamic override and public import path
