+++ openhands/integrations/github/service/graphql.py
@@ -1,0 +1,126 @@
+import json
+from typing import Any
+
+import httpx
+
+from openhands.core.logger import openhands_logger as logger
+from openhands.integrations.github.queries import (
+    suggested_task_issue_graphql_query,
+    suggested_task_pr_graphql_query,
+)
+from openhands.integrations.github.service._base import GitHubMixinBase
+from openhands.integrations.service_types import (
+    ProviderType,
+    SuggestedTask,
+    TaskType,
+    UnknownException,
+)
+
+
+class GitHubGraphQLMixin(GitHubMixinBase):
+    async def execute_graphql_query(
+        self, query: str, variables: dict[str, Any]
+    ) -> dict[str, Any]:
+        try:
+            async with httpx.AsyncClient() as client:
+                github_headers = await self._get_github_headers()
+                response = await client.post(
+                    f'{self.BASE_URL}/graphql',
+                    headers=github_headers,
+                    json={'query': query, 'variables': variables},
+                )
+                response.raise_for_status()
+
+                result = response.json()
+                if 'errors' in result:
+                    raise UnknownException(
+                        f'GraphQL query error: {json.dumps(result["errors"])}'
+                    )
+
+                return dict(result)
+
+        except httpx.HTTPStatusError as e:
+            raise self.handle_http_status_error(e)
+        except httpx.HTTPError as e:
+            raise self.handle_http_error(e)
+
+    async def get_suggested_tasks(self) -> list[SuggestedTask]:
+        user = await self.get_user()
+        login = user.login
+        tasks: list[SuggestedTask] = []
+        variables = {'login': login}
+
+        try:
+            pr_response = await self.execute_graphql_query(
+                suggested_task_pr_graphql_query, variables
+            )
+            pr_data = pr_response['data']['user']
+
+            for pr in pr_data['pullRequests']['nodes']:
+                repo_name = pr['repository']['nameWithOwner']
+                task_type = TaskType.OPEN_PR
+
+                if pr['mergeable'] == 'CONFLICTING':
+                    task_type = TaskType.MERGE_CONFLICTS
+                elif (
+                    pr['commits']['nodes']
+                    and pr['commits']['nodes'][0]['commit']['statusCheckRollup']
+                    and pr['commits']['nodes'][0]['commit']['statusCheckRollup'][
+                        'state'
+                    ]
+                    == 'FAILURE'
+                ):
+                    task_type = TaskType.FAILING_CHECKS
+                elif any(
+                    review['state'] in ['CHANGES_REQUESTED', 'COMMENTED']
+                    for review in pr['reviews']['nodes']
+                ):
+                    task_type = TaskType.UNRESOLVED_COMMENTS
+
+                if task_type != TaskType.OPEN_PR:
+                    tasks.append(
+                        SuggestedTask(
+                            git_provider=ProviderType.GITHUB,
+                            task_type=task_type,
+                            repo=repo_name,
+                            issue_number=pr['number'],
+                            title=pr['title'],
+                        )
+                    )
+        except Exception as e:
+            logger.info(
+                f'Error fetching suggested task for PRs: {e}',
+                extra={
+                    'signal': 'github_suggested_tasks',
+                    'user_id': self.external_auth_id,
+                },
+            )
+
+        try:
+            issue_response = await self.execute_graphql_query(
+                suggested_task_issue_graphql_query, variables
+            )
+            issue_data = issue_response['data']['user']
+
+            for issue in issue_data['issues']['nodes']:
+                repo_name = issue['repository']['nameWithOwner']
+                tasks.append(
+                    SuggestedTask(
+                        git_provider=ProviderType.GITHUB,
+                        task_type=TaskType.OPEN_ISSUE,
+                        repo=repo_name,
+                        issue_number=issue['number'],
+                        title=issue['title'],
+                    )
+                )
+            return tasks
+        except Exception as e:
+            logger.info(
+                f'Error fetching suggested task for issues: {e}',
+                extra={
+                    'signal': 'github_suggested_tasks',
+                    'user_id': self.external_auth_id,
+                },
+            )
+
+        return tasks
